product

+++ src/main/java/com/sala/java/school/phoneshope/Exception/ResourceNotFoundException.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 
 public class ResourceNotFoundException extends ApiException {
 
-	public ResourceNotFoundException(String resouceName, Integer id) {
+	public ResourceNotFoundException(String resouceName, Long id) {
 		super(HttpStatus.NOT_FOUND, String.format("%s with id = %d not found", resouceName, id));
 	}
