+++ src/main/java/com/sala/java/school/phoneshope/repository/ModelRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.sala.java.school.phoneshope.entity.Model;
 
 @Repository
-public interface ModelRepository extends JpaRepository<Model, Integer>{
+public interface ModelRepository extends JpaRepository<Model, Long>{
 	List<Model> findByBrandId(Integer brandId);
 }
