+++ src/main/java/com/sala/java/school/phoneshope/service/BrandService.java
@@ -10,9 +10,9 @@ import com.sala.java.school.phoneshope.entity.Brand;
 public interface BrandService {
 	Brand create(Brand brand);
 
-	Brand getById(Integer id);
+	Brand getById(Long id);
 
-	Brand update(Integer id, Brand brandUpdate);
+	Brand update(Long id, Brand brandUpdate);
 
 	List<Brand> getBrands(String name);
