+++ src/main/java/com/sala/java/school/phoneshope/service/Impl/BrandServiceImpl.java
@@ -18,6 +18,7 @@ import com.sala.java.school.phoneshope.service.Util.PageUtil;
 
 @Service
 public class BrandServiceImpl implements BrandService {
+	
 	@Autowired
 	private BrandRepository brandRepository;
 
@@ -27,7 +28,7 @@ public class BrandServiceImpl implements BrandService {
 	}
 
 	@Override
-	public Brand getById(Integer id) {
+	public Brand getById(Long id) {
 
 //		Optional<Brand> brandOptional = brandRepository.findById(id);
 //		if(brandOptional.isPresent()) {
@@ -43,7 +44,7 @@ public class BrandServiceImpl implements BrandService {
 	}
 
 	@Override
-	public Brand update(Integer id, Brand brandUpdate) {
+	public Brand update(Long id, Brand brandUpdate) {
 		Brand brand = getById(id);
 		brand.setName(brandUpdate.getName()); // @TODO improve update
 		return brandRepository.save(brand);
